pyproject.toml
[project]
name = "denys-portfolio"
version = "1.0.0"
description = "Personal portfolio & blog — FastAPI + DDD"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "pydantic-settings>=2.5",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1",
    "markdown>=3.5",
    "nh3>=0.2",
    "alembic>=1.13",
    "httpx>=0.27",
    "dependency-injector>=4.41",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "mypy>=1.11",
    "ruff>=0.6",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.10"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
